rework application build script

+++ src/OpcUaProjectBuilder/ProjectTemplate/build.sh
@@ -15,19 +15,15 @@ CMAKE_GENERATOR_LOCAL=-G"Eclipse CDT4 - Unix Makefiles"
 
 usage()
 {
-   echo "build.sh <Command>"
+   echo "build.sh (info | local | deb | rpm | tst | clean)"
    echo ""
-   echo "Commands:"
-   echo "  info" 
-   echo "    create version and dependency files"
-   echo "  local [renew]"
-   echo "    create or renew local build and install in folder ${HOME}/install"
-   echo "  deb [renew]"
-   echo "    create or renew deb package"
-   echo "  rpm"
-   echo "    create or renew rpm package"
-   echo "  clean"
-   echo "    delete all build directories"
+   echo "  info  - create version and dependency files"
+   echo "  local - create local build and install in folder ${HOME}/install"
+   echo "  deb   - create deb package"
+   echo "  rpm   - create rpm package"
+   echo "  tst   - build unit application"
+   echo "  clean - delete all build directories"
+
 }
 
 
@@ -38,7 +34,7 @@ usage()
 #
 # -----------------------------------------------------------------------------
 # -----------------------------------------------------------------------------
-version_info()
+build_info()
 {
     echo "version info"
 
@@ -48,7 +44,7 @@ version_info()
     cmake -DCREATE_INFO_FILES=1 ../src/
 }
 
-version_info_clean()
+build_info_clean()
 {
     rm -rf build_info
 }
@@ -66,21 +62,33 @@ build_local()
     echo "build local start"
 
     # build local directory
-    rm -rf build_local
-    mkdir build_local
+    if [ ! -d "build_local" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_local
+        mkdir build_local
+    else
+        BUILD_FIRST=0
+    fi
     cd build_local
 
+
     # build local
-    : ${OPCUASTACK_INSTALL_PREFIX:=${HOME}/install}
-    set -x
-    cmake ../src \
-          -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
-          "${CMAKE_GENERATOR_LOCAL}" 
-    set +x
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+	: ${OPCUASTACK_INSTALL_PREFIX:=${HOME}/install}
+	set -x
+	cmake ../src \
+	    -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
+	    "${CMAKE_GENERATOR_LOCAL}" 
+	set +x
+	if [ $? -ne 0 ] ;
+	then
+	    echo "cmake error"
+	    return $?
+	fi
+    else
+        cmake .
     fi
 
     # install local
@@ -93,40 +101,6 @@ build_local()
 
     return 0
 }
-
-build_local_renew()
-{
-    echo "renew local start"
-
-    # build local directory
-    if [ ! -d "build_local" ] ; 
-    then
-        echo "error - directory build_local do not exist"
-        return 1;
-    fi
-    cd build_local
-
-    # build local
-    set -x
-    cmake .
-    set +x
-    if [ $? -ne 0 ] ;
-    then
-        echo "cmake error"
-        return $?
-    fi
-
-    # install local
-    make DESTDIR="${HOME}/install" install
-     if [ $? -ne 0 ] ;
-    then
-        echo "make install error"
-        return $?
-    fi
-
-    return 0
-}
-
 
 build_local_clean()
 {
@@ -155,22 +129,35 @@ build_deb()
     fi
 
     # build package directory
-    rm -rf build_deb
-    mkdir build_deb
+    if [ ! -d "build_deb" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_deb
+        mkdir build_deb
+    else
+        BUILD_FIRST=0
+    fi
     cd build_deb
 
+
     # build package
-   : ${OPCUASTACK_INSTALL_PREFIX:=/}
-    cmake ../src \
-        -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
-        "${CMAKE_GENERATOR_LOCAL}" \
-        "-DCPACK_BINARY_DEB=1" \
-        "-DCPACK_BINARY_RPM=0" 
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+	: ${OPCUASTACK_INSTALL_PREFIX:=/}
+	cmake ../src \
+	    -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
+	    "${CMAKE_GENERATOR_LOCAL}" \
+	    "-DCPACK_BINARY_DEB=1" \
+	    "-DCPACK_BINARY_RPM=0" 
+	if [ $? -ne 0 ] ;
+	then
+	    echo "cmake error"
+	    return $?
+	fi
+    else
+        cmake .
     fi
+
     make package
     if [ $? -ne 0 ] ;
     then
@@ -180,37 +167,6 @@ build_deb()
 
     return 0
 }
-
-
-build_deb_renew()
-{
-    echo "renew deb package start"
-
-    # build package directory
-    if [ ! -d "build_deb" ] ; 
-    then
-        echo "error - directory build_deb do not exist"
-        return 1;
-    fi
-    cd build_deb
-
-    # build package
-    cmake .
-    if [ $? -ne 0 ] ;
-    then
-        echo "cmake error"
-        return $?
-    fi
-    make package
-    if [ $? -ne 0 ] ;
-    then
-        echo "make package error"
-        return $?
-    fi
-
-    return 0
-}
-
 
 build_deb_clean()
 {
@@ -240,63 +196,35 @@ build_rpm()
     fi
     
     # build package directory
-    rm -rf build_rpm
-    mkdir build_rpm
-    cd build_rpm
-
-    # build package
-    echo "${CPACK_BINARY}"
-    : ${OPCUASTACK_INSTALL_PREFIX:=/}
-    cmake ../src \
-        -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
-        "${CMAKE_GENERATOR_LOCAL}" \
-        "-DCPACK_BINARY_DEB=0" \
-        "-DCPACK_BINARY_RPM=1"
-    if [ $? -ne 0 ] ;
+    if [ ! -d "build_rpm" ] ;
     then
-        echo "cmake error"
-        return $?
-    fi
-    make package
-    if [ $? -ne 0 ] ;
-    then
-        echo "make package error"
-        return $?
-    fi
-
-    return 0
-}
-
-
-build_rpm_renew()
-{
-    echo "build rpm package start"
-
-    # check if rpm package to be produced
-    if which rpmbuild > /dev/null ;
-    then
-        echo "build rpm package enable"
+        BUILD_FIRST=1
+        rm -rf build_rpm
+        mkdir build_rpm
     else
-        echo "build rpm package disable"
-        return -1
-    fi
-    
-    # build package directory
-    if [ ! -d "build_rpm" ] ; 
-    then
-        echo "error - directory build_rpm do not exist"
-        return 1;
+        BUILD_FIRST=0
     fi
     cd build_rpm
 
+
     # build package
-    echo "${CPACK_BINARY}"
-    cmake .
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+	: ${OPCUASTACK_INSTALL_PREFIX:=/}
+	cmake ../src \
+	    -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
+	    "${CMAKE_GENERATOR_LOCAL}" \
+	    "-DCPACK_BINARY_DEB=0" \
+	    "-DCPACK_BINARY_RPM=1"
+	if [ $? -ne 0 ] ;
+	then
+	    echo "cmake error"
+	    return $?
+	fi
+    else
+        cmake .
     fi
+
     make package
     if [ $? -ne 0 ] ;
     then
@@ -306,7 +234,6 @@ build_rpm_renew()
 
     return 0
 }
-
 
 
 build_rpm_clean()
@@ -316,26 +243,46 @@ build_rpm_clean()
 
 
 
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
+#
+# build unittest
+#
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 build_tst()
 {    
     echo "build tst start"
 
     # build tst directory
-    rm -rf build_tst
-    mkdir build_tst
+    if [ ! -d "build_tst" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_tst
+        mkdir build_tst
+    else
+        BUILD_FIRST=0
+    fi
     cd build_tst
 
+
     # build tst
-    : ${OPCUASTACK_INSTALL_PREFIX:=${HOME}/install}
-    cmake ../tst \
-        -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
-  	"${CMAKE_GENERATOR_LOCAL}" \
-	-DOPCUASTACK_INSTALL_PREFIX="${HOME}/install"
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+        : ${OPCUASTACK_INSTALL_PREFIX:=${HOME}/install}
+        cmake ../tst \
+            -DOPCUASTACK_INSTALL_PREFIX=${OPCUASTACK_INSTALL_PREFIX} \
+  	    "${CMAKE_GENERATOR_LOCAL}" \
+	    -DOPCUASTACK_INSTALL_PREFIX="${HOME}/install"
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+    else 
+        cmake .
     fi
+
     make 
      if [ $? -ne 0 ] ;
     then
@@ -370,59 +317,36 @@ clean()
 #
 # -----------------------------------------------------------------------------
 # -----------------------------------------------------------------------------
-if [ $# -eq 1 ] ; 
+if [ $# -ne 1 ] ; 
 then
-    RENEW=0
-elif [ $# -eq 2 ] ;
-then
-    if [ "$2" = "renew" ] ;
-    then
-        RENEW=1
-    else
-        usage
-        exit 1
-    fi
-else
     usage
     exit 1
 fi 
 
 if [ "$1" = "info" ] ;
 then
-    version_info
+    build_info
     exit $?
 elif [ "$1" = "clean" ] ;
 then 
     clean 
 elif [ "$1" = "local" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_local_renew
-    else
-        build_local
-    fi 
+    build_local
     exit $?
 elif [ "$1" = "deb" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_deb_renew
-    else
-        build_deb 
-    fi
+    build_deb 
     exit $?
 elif [ "$1" = "rpm" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_rpm_renew
-    else 
-        build_rpm
-    fi 
+    build_rpm
+    exit $?
+elif [ "$1" = "tst" ] ;
+then 
+    build_tst
     exit $?
 else
     usage
+    exit 1
 fi
-
-exit 0
